+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- defined all state in MainRadioControlStateMachineState as in datasheet page 93 
+
 ### Removed
 - compatibility with `python3.7` & `python3.8`